We analyze with a Cox model the data contained in 'dta' to compare 2 groups (for example treatment)

We analyze with a Cox model the data contained in 'dta' to compare 2 groups (for example treatment)

Usage

ana_mult_2gr(sim)

Arguments

sim

the list containing simulations data set and parameters created by using nlisttrial

Value

the function returns a list containing a dataset containing, for each estimation approach, each scenario and each repetition:

the log Hazard Ratio

the Hazard Ratio

the statistics of the Wald test

the p-value associated with the Wald test

the success in unilateral (0 = no, yes = 1) if p-value < 0.05 & HR < 1

the failure (0 = no, yes = 1) if p-value < 0.05 & HR > 1

the success in bilateral (0 = no, yes = 1) if p-value < 0.05 And a dataset containing simulations parameters