We simulate different trials with possibly different parameters specified in arguments for each scenario

We simulate different trials with possibly different parameters specified in arguments for each scenario

Usage

nlisttrial(
  scenario = 1,
  nrepeat = 1,
  nbp = 1000,
  tacc = 60,
  tfu = 60,
  lambda0 = 0.03,
  HR = 1,
  lambda_pdv1 = 0.01,
  lambda_pdv2 = 0.02,
  t_pdv2 = 36,
  HR_pdv = 1,
  pinsee0 = 0.95,
  pinsee1 = 0.95,
  seed = 2023
)

Arguments

scenario

: the number of the scenarios (default 1 scenario)

nrepeat

: a vector containing the number of repetitions for each scenario (default 1 repetition)

nbp

: a vector containing the number of patients for each scenario (defaut 1,000 patients)

tacc

: a vector containing the accrual duration in months for each scenario (default 60 months)

tfu

: a vector containing the follow-up duration in months for each scenario (default 60 months)

lambda0

: a vector containing the rate of the exponential distribution for each scenario (default 0.03)

HR

: a vector containing the Hazard Ratio for survival between group 0 and group 1 for each scenario (default 1)

lambda_pdv1

: a vector containing the rate of lost to follow-up for patients until time 't_pdv2' for each scenario (default 0.01)

lambda_pdv2

: a vector containing the rate of lost to follow-up for patients after time 't_pdv2' for each scenario (default 0.02)

t_pdv2

: a vector containing the time at which rate of lost to follow-up changes for each scenario (default 36)

HR_pdv

: a vector containing the Hazard ratio for lost to follow-up between group 0 and group 1 for each scenario (default 1)

pinsee0

: a vector containing the administrative file completeness (insee) in group 0 for each scenario (default 0.95)

pinsee1

: a vector containing the administrative file completeness (insee) in group 1 for each scenario (default 0.95)

seed

: seed used to generate each scenario

Value

the function returns a simulation dataset with all the scenarios